moved helper class out of test file and into TestDoubles subdir to satisfy phpcs
