+++ tests/Command/CheckS3PermissionsCommandTest.php
