+++ tests/Command/UpdateFrontendCommandTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Tests\Command;
 
-use App\Command\UpdateFrontendCommand;
+use App\Tests\Command\TestDoubles\TestableUpdateFrontendCommand;
 use App\Service\Archive;
 use App\Service\Config;
 use App\Service\Fetch;
@@ -109,37 +109,5 @@ final class UpdateFrontendCommandTest extends KernelTestCase
             'Frontend updated successfully from staging build!',
             $this->commandTester->getDisplay()
         );
-    }
-}
-
-class TestableUpdateFrontendCommand extends UpdateFrontendCommand
-{
-    private ?string $currentVersion = null;
-    private array $distributions = [];
-
-    public function setTestData(?string $currentVersion, array $distributions): void
-    {
-        $this->currentVersion = $currentVersion;
-        $this->distributions = $distributions;
-    }
-
-    protected function downloadAndExtractAllArchives(string $environment): ?string
-    {
-        return $this->currentVersion;
-    }
-
-    protected function listDistributions(string $environment): array
-    {
-        return $this->distributions;
-    }
-
-    protected function copyAssetsIntoPublicDirectory(string $distributionPath): void
-    {
-        // no-op
-    }
-
-    protected function activateVersion(string $distributionPath): void
-    {
-        // no-op
     }
 }
